airflow: Use the existing labels for kubernetes and spark operators
We had attempted to add a custom 'operator' label to the pods generated
by the KubernetesPodOperator, but it seems that they weren't being added
and so the networkpolicy was not applied.

However, the airflow providers for the kubernetes pod and spark job
operators do already add labels that set "kubernetes_pod_operator":
"True" and "spark_kubernetes_operator": "True", respectively.

Therefore, we can use these labels as selectors for the networkpolicy
that allows access to the k8s API.

Bug: T383430
Bug: T318712
Change-Id: Ifc468315ecfedf7e0f3e41d20b4a63aa44c41fe9