airflow: Allow the scheduler to patch existing pods
We have found that when a scheduler restarts (for whatever reason) it
attempts to 'adopt' the currently running task pods that are found in
the namespace.

In order to do tihs is tries to patch the pods, but it fails with a 403
error message.

This commit adds the rights for the scheduler to patch existing pods.

Bug: T380621
Change-Id: Id02a7258ebff5821c3a5a77a624f12fbb97d7e97