airflow: refactor/DRY the volume/volumeMounts accross containers
One of the issues we've been having as maintainers of the airflow chart
was the scattering and hardcoding of the volumes and volumeMounts
accross the values, templates and helpers.

I've tried to clean that up the following way.

In values.yaml:
- app.volumes/volumeMounts manages all the airflow configuration
  files
- scheduler.volumes/volumeMounts: points to app.volume/VolumeMounts
- worker.volumes/volumeMounts contains groups of volume(mount)s that can
  be picked or left un-used, depending on what pod_template will be in
  used (see later)
- kerberos.volumes/volumeMounts contains 2 of such groups:
  - base: contains the token and configuration file
  - keytab: contains the keytab (which should only be used by the
    airflow-kerberos token rewnewer deployment)

In _helpers.tpl:
- I defined macros managing the rendering of volumes/volumeMounts
  blocks.
- Some macros can take a `profiles` parameter, toggling on/off
  which groups should be used. Ex: hadoop/airflow/spark, etc

This not only cleans up the poor state of the chart, but also opens
up the way to define another pod_template for pods created by the
KubernetesPodOperator that need to have access to the, eg, hadoop
config files.

We also fix the fact that the keytab was mounted in all compomnents,
when it should only be mounted in the airflow-kerberos pod.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: I44e38a586976d7a025be51afc652f056a10482dc
Bug: T383430