airflow: bypass DNS resolution for the PG URI
We are observing frequent DNS "Temporarily resolution failure"
occurences throughout all airflow instances, specifically when some
python code is resolving a PG hostname. We *think* that we have tracked
it down to some known bug in libpq.

As the clusterIP behind the PoolerRW service is stable, we can safely
replace the service name by the cluster IP. It's a bit of a hack, but it
should help us stabilize the scheduler.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T383651
Change-Id: I32de57314f5def317d6acdf0025b0604f8b625cf