shellbox-video: 3 codfw replicas on 8.1 (change 1/4)
Pilot with 3 replicas on 8.1, corresponding to ~ 9% of traffic.

Unlike other shellbox instances migrated so far, video will in practice
only receive traffic in the current primary DC (codfw).

Why 3? shellbox-video is larger (30 replicas) and sees lower request
rates than the other variants, so a single replica would likely require
a rather long soak before serving a single request. As a compromise, 3
replicas corresponds to ~ 9% of requests, which is similar to the pilot
size we've used elsewhere.

Given that -video is also fairly low risk, this seems warranted.

Bug: T377038
Change-Id: I07b87366d84b261f54b13a033cdb347e04c8fa23