airflow: remove useless separator in pod spec confusing the config checksum computation
I found out that updates made to the airflow-kubernetes-pod-templates Configmap did not
change the `checksum/configuration` annotation included in the 3 airflow Deployments
(webserver, scheduler and kerberos).

The helper in charge of computing that checksum renders the whole
`configmap.yaml` file, splits it on `---\n`, adds the content of `.data`
to a list and checksums the whole list.

The issue is: splitting over the `---\n` separator within the pod
specs themselves, produced badly indented spec, which were deemed
invalid and _silently_ ignored. Yay YAML engineering.

Long story short, we drop these separators (useless in themselves, as
there's nothing to separate in a single-document spec), which restores
the appropriate behavior: updating the pod templates will update the
Deployment checksums, which will restart the pods and pull the latest
version of the config files.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T384275
Change-Id: Iedaa8d64ff1de1485619ba82ec527e8f19a5e5f8