CI: Ensure admin checks don't run unnecessary template calls
We used to call helmfile template on all admin_ng environments
regardless of that the user chose to run. This change ensures the
fixtures of AdminAssets (list of environments in that case) is always
properly filtered.

Allow admin_lint to run for certain environments only

Fix ERB.new() call to no longer emit deprecation warnings

Change-Id: I80800e0d2ca359f6b0da827fe680045c115a4bee