mediwiki-dumps-legacy: Create helmfile deployment of a suspended job
We wish to start using mediwiki based depeloyments on the dse-k8s-eqiad
cluster to perform dumps. This helmfile deployment will be responsible
for creating a number of supporting kubernetes objects in the
mediwiki-dumps-legacy namespace.

One of them will be a suspended job template, which will also be
permitted to mount a PersistentVolume.

Bug: T352650
Change-Id: I4a16c0104083cda3c6499b5c78ef0c69d087e3f9