mw-(api-ext|web): scale next to 10% of main
As an initial scale-up step to unblock the small-% traffic ramp planned
for the week of 2025-01-27, scale both next releases to 10% of main.

Rationale: in week 1, we plan to get to ~ 5% of cookie-enrollable
traffic, in steps of 1, 2, and 5%.

If we assume that (1) all traffic to these deployments is enrollable
(optimistic) and (2) enrolled population is 1:1 with traffic migrated
(unlikely), then this is 2x what we "anticipate" needing in week 1.

While both are questionable, we should get a sense of this empirically
early in the week (1-2% enrolled) and can adjust accordingly.

Note: Since these are rather small bumps compared to aggregate cluster
capacity, no attempt to reclaim capacity from main is planned. This also
simplifies rollbacks.

Bug: T383845
Change-Id: Ib0a29b7f20b1651702820a2c67dea92c13f3f394