mw-(api-ext|web): temporarily increase maxUnavailable
The combination of small-ish deployments size with "rare" image means
that it is also rare to have large image layers cached locally. This
makes deployments far too slow at 6% maxUnavailable (inherited).

Instead, since we're also very over-provisioned at the moment, raise
maxUnavailable to 20%, with a TODO to tune down as the deployment grows.

Bug: T383845
Change-Id: I9cddfcde982d3b0f052ae43461e61afdf597ac69