dse-k8s-eqiad: deploy the sidecar job controller
We are planning to run airlfow tasks pods with an envoy sidecar, which
would allow the jobs to reach out to internal systems using the service
mesh.

However, we found out that when such a pod's main container command
finihses, the envoy sidecar stays running, causing the Pod to get into a
NotReady state, and is left running forever.

This is incidently a solved issue, as Job resources created by the
mediawiki chart deployments are created with envoy sidecars (and even
possibly more).

We thus deploy the sidecar controller, in charge of deleting pods with
their main container in a Finished state, by stopping all sidecars.

For now, we only assign the airflow-test-k8s namespace to the
controller.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T384329
Change-Id: I5f04c3cf71f05b01eaf2a3291274875d31f0445f