custom_deploy.d: rework dse-k8s-eqiad's istio config
For some reason istiod is not picking up the seccomp profile
settings/Annotation in its pods. The only difference that I see
from main/aux (that have the setting applied) is related to
component vs additionalComponents.
The istioctl manifest diff seems to agree that the change is
confined to these two things, and it is good to keep our configs
inline without diverging too much.

Change-Id: I91f3e105590991fc6f3f455dc73fdf8456e1d20f