Scale down mw-content-history-reconcile-enrich for nominal events intake
On 1108801 we temporarily bumped tasks managers from 2 to 20. On this patch we revert that to the nominal value.

We do keep the memory bump from 1108801 (from 1.5GB to 3.0GB) as the task managers have not died on us since that change, while before they did die unexpectedly a couple times.

Bug: T382953
Bug: T358375
Change-Id: If7a95523786edee39a8d84f5eb6cc4f798021c82