mediawiki: introduce feature flags
Much like we do in configuring MediaWiki in production, we use the
"servergroup" variable to determine a series of feature flags
influencing the behaviour of the mediawiki chart, without resorting
to long conditionals to do so.

Please note: we can still change all of these trait names and
feature flag names and / or expand them. I'd avoid bikeshedding on
this patch though as it's somewhat urgent we establish a mechanism
to protect against proliferation of complex conditionals in the chart.

Change-Id: I669847bd81609e406e21caa84513c9884bb45b88