jobqueue: bump ThumbnailRender concurrency
ThumbnailRender concurrency is currently quite low compared to
historical values as a result of various thumbor bugs and issues in
the past. We're in a relatively good place with Thumbor stability at
the moment, so increase this value a little to start. Before we
started adjusting this number the concurrency was 20, so this is still
a fairly conservative move.

Bug: T385273
Change-Id: Ie16cb64210219a5824b9c6d2ac1e602ca545b713