mw-(api-ext|web): scale next to 15% of main
So far, we've seen that the amount of traffic landing on the next
releases:
1. scales roughly linearly with the enrollment fraction
2. remains significantly less than one would expect if one simply
   multiplied the enrollment fraction by the amount landing on main,
   e.g., despite being enrolled at 10% we're seeing less than 5% of
   main's traffic levels.

Taken together with the current PHP-FPM worker utilization, I believe
that scaling up by 1/2 (i.e., to 15% of main) should be sufficient to at
least move to 25% enrollment, if not more.

I've not yet started to "swap" capacity from main, as we're still at a
point where current replica counts are not prohibitive in terms of
cluster capacity.

Bug: T383845
Change-Id: I7960733a444d1fc00eb0079e9ebbc8f3359b23bb