Add the networkpolicy feature flag
This will allow us to declare a single networkpolicy
in namespaces where we have a lot of releases,
like mwcron and mwscript.

Once we've added an appropriate second chart, we'll set the flag
for networkpolicy to "false" in the cronjobs and scripts deployments,
and add a further release of a "mediawiki-common" chart to those
namespaces.

Change-Id: I4a352d4afe2d139139450e3995c5f836aef5283f