mediawiki-common: introduce chart
In some namespaces, we want to execute Job and CronJobs with different
release names, in large numbers. When we do so, a lot of resources
that could be shared between such releases are duplicated, like
configmaps, secrets and networkpolicies.

This can create scalability issues (see the case in which scheduling
1000 jobs cause a global calico outage because of the amount of
networkpolicies involved), so this chart has the goal to overcome that.

When you're in such a situation, you should add a release to your
helmfile using the mediawiki-common chart, and disable the addition of
networkpolicies in the feature flags of the mediawiki chart.

Change-Id: I0239868b21d23cac39e60e3fd4269b7e79a5b4d1