Add a mediawiki-common release to mw-script
When running a script, a full mediawiki release is deployed, which means
a lot of resources that don't change between script executions. This on
one side bloats the amount of objects we have on the k8s api servers,
and on the other causes too many networkpolicies to be defined, and that
causes outages as calico can't keep up with the amount of different
networkpolicies to manage.

To avoid this effect and still allowing people to run script on many
wikis in tight sequence, we have created a "mediawiki-common" chart that
for now just holds the networkpolicy with a broader selector than usual,
and at the same time we stop declaring the networkpolicy for the
mw-script executions. This should allow us to create more script
instances without risking to bring down the kubernetes networking.

Change-Id: I101f0db0cb6d59fc87210293a91f5efbefc69233