knserve-inference: add seccompProfile to the pod security context
To allow the PSS k8s restricted policy to be applied, all the containers
in the pod need to run seccompProfile to RuntimeDefault or Localhost.

This change adds a security context for the pod (hence for all
the containers) that does exactly that. This setting will take
care of containers like istio, that don't have a way to set
seccompProfile (at least not in our version).

Bug: T369493
Change-Id: Ic6c9506d2e868d7a1c7a5fef4c541388bc51beee