mediawiki: Fix default-merging logic in _site_helpers.tpl
This line applies multiple layers of defaults, falling back to the more
and more generic layers when keys are missing in the more specific
maps.

Previously we used `merge` to do that, but that has the flaw that it
doesn't semantically distinguish between zero values and missing ones:
it treats the map {foo: false} as the same as the map {}. Thus when a
key has the default value true, it can never be overridden to false.

Here we replace that with `mergeOverwrite`, which does treat zero values
different from unset ones. Note that the docs[1][2] don't specify this
difference -- in fact the behavior around zero values is undocumented
for both merge and mergeOverwrite, and seems to be an implementation
detail(?) which we hope won't change in future versions(?!).

Informally, this workaround is variously reported to be unsuccessful[3]
and successful[4], so the implementation may already have changed once.
Testing on our infra today, we find it does work.

What the docs do specify is that mergeOverwrite takes its arguments in
the opposite order from merge, so we reverse them here, and that you'd
better use a deep copy if you don't want shared references, so we do so
here. (And in fact, testing without the calls to deepCopy does result in
smearing values out across our various vhosts as each mergeOverwrite
points them all to the same underlying mapping.)

This change is not a no-op; there ARE Apache config diffs associated
with it. Due to this bug we were ignoring a whole bunch of instances of
`legacy_rewrites: false` in httpd.yaml; they had no effect since it's
true in tpl_defaults. With this fix, those rewrites are actually
disabled again (as they would have been pre-mw-on-k8s). I believe that
change to be safe to roll out, and I'm not aware of anyone noticing that
behavior externally, but it would have been visible.

[1]: https://helm.sh/docs/chart_template_guide/function_list/#merge-mustmerge
[2]: https://helm.sh/docs/chart_template_guide/function_list/#mergeoverwrite-mustmergeoverwrite
[3]: https://github.com/helm/helm/issues/5238#issuecomment-650073864
[4]: https://github.com/helm/helm/issues/5238#issuecomment-2454093900

Bug: T385228
Change-Id: Ia6ab2da984a659e283b946f25f43d337ed67dc40