services: double the capacity for Kartotherian in production
The Kartotherian bare metal servers are 6 for each DC, providing
48 vcores each. Nodejs is set to run with nproc workers for each
server, since Kartotherian is memory hungry (high concurrency and
high cpu usage). With a single kubernetes host pooled in the LVS
we handle traffic just fine, but when we add the second one the cpu
usage across containers doubles and the latency follows as well
(p99 jumps 10 folds). We'll probably won't need to match the 48x6
capacity in each DC, but we'll surely need to scale up horizontally
to spread the load.

Each pod is currently requiring 5000 millicores of cpu, so doubling
the current capacity is a sizeable request. I am reasonably sure that
it will not be enough, but it is surely a good starting point to
see if scaling horizontally helps handling more load.

Bug: T386926
Change-Id: Ide6f4d4ef57971618402f62a4ebe4c22d909f8a2