services: update Kartotherian's replicas to 20
The recent bump from 5 to 10 helped a lot with high latency
and average CPU usage, so the good news is that Kartotherian
can scale horizontally in a decent way. I would keep the current
setup for single pods, and keep increasing their number replicas
instead of trying to bump their CPU requirements vertically.

You can see more details in the task, but with 3 k8s workers
pooled in the LVS at half weight of the bare metals we are
able to handle ~150 rps, and we area already again reaching
high numbers of CPU usage (some throttling already happening).

My proposal is to double the capacity, from 10 to 20, asking
10x5000m + 10x1GiB more capacity in both eqiad and codfw.
The new capacity should hopefully get us close to absorb all
the traffic that bare metal handles.

Raise also the Resource Quota settings to allow the requested
capacity to be spawned (plus an extra headroom for scaling if
needed).

Bug: T386926
Change-Id: If52193c7167abf768245342161c364686a71417f