kserve-inference: move pod security settings for seccomp to staging only
For some misterious reasons, the first deployment to prod with
the seccomp profile enabled (at the pod level) caused some istio-proxy
issues, ending up with loss of connectivity for the pod etc..

While we investigate the issue, it is best to confine these updates
only to staging, so prod's deployments can flow without any risk.

Add also the seccomp settings to the transformer pods, something
that I forgot during the first pass (applied only to predictors).

Bug: T369493
Change-Id: I112b3bc22dff19f8de4446f85058e42cde8e007f