cirrus-streaming-updater: scale up the consumer-search job
The elasticsearch sink appears to be the bottleneck.
The assumption is that having only two clients writing to eqiad is not
enough because the bulk indexing requests are slightly slower there.

Add one more task manager to further increase the parallisism of the
sink moving from 2 to 3.

Keep a parallelism of 2 for cloudelastic which appears to be running
fine so far.

Bug: T386935
Change-Id: I856f1236094889d4362327dc3b2ddd15a1696bcb