airflow-test-k8s: temporarily mimic airflow-analytics
We are going to temporarily use airflow-test-k8s as an integration
testing ground for the DAGs we migrate over from an-launcher1002 to
airflow-main.

As such, git-sync will continuously pull from a feature branch, and the
airflow instance should be given the same configuration files, network
policies and general behavior than airflow-analytics, except that it
the DAGs it is running should live under `airflow_dags/main`.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T386282
Change-Id: Ie16091d00182d43e4dcac4b1e370ceb496fba815