kserve-inference: remove the need for the kserve container's securityContext
Knative Serving is responsible for this part, namely injecting
the security features in its managed containers.

Tested in ml-staging-codfw where we have the patched version of Knative
that takes care of what described above.

This should be safe to deploy to production, we don't have any current
security restrictions and the PSP policies take care of injecting
what's needed.

Bug: T369493
Change-Id: I05412ba509c9f1ec1149f5756fb4d3abc14a022f