benthos: add input/output config to chart
What:
Allow for configuring benthos input, output and metrics.
Also switch the default input and output from kafka to kafka_franz, as
that one is almost always what we want (better performance and logging).

Why:
We need the chart to be a bit less bare-bones / more configurable in
preparation for moving MW accesslog metrics to k8s.

Bug: T371214
Change-Id: Ibd5941cb2375f7dc86461e27ff0ee11e27eb9d5d