benthos-mw-accesslog-metrics: create deployment
What:
Create a deployment in k8s-aux that replicates the functionality of the
mw_accesslog_metrics Benthos instance on centrallog:
- consumes the mediawiki.httpd.accesslog Kafka topic
- converts data from the log into metrics
- exposes the metrics with histogram buckets

Why:
Migration to k8s.

After this runs, I will:
1. check the existing dashboards to make sure they're not getting
   duplicate metrics
2. check that these metrics match the bare metal instance
3. edit dashboards to use these metrics, including dealing with now
   having multiple benthos instances
4. check resource consumption and adjust resources+limits
5. turn down the bare metal service

I will also turn this up in aux-k8s-codfw once we have it, for symmetry
and ease of failover, but IIUC only the eqiad Kafka actually receives
messages, so this should be sufficient.

Change-Id: I8da306c5f1092b18aa2339150846a85f18ef6a79