mw-(api-ext|web): scale next to 40% of main
Based on current trends, we expect 100% of the population of enrollable
clients to represent ~ 30% of total traffic. This scales next to 40% of
main's size, which leaves some additional buffer.

This must be merged and applied before the enrollment fraction change in
I847a410c90e62c8f4eedb51da47b1764e24ded12 is deployed.

This does not yet scale down main to free up capacity for next, since:
1. this increment is "only" ~ 400 CPU-s/s per cluster, which fits well
   within current headroom
2. doing so adds complexity to upcoming work to validate single-DC
   serving capacity ahead of the March switchover.

Also, now that the next deployments have grown significantly, cut
maxUnavailable in half (though not yet reverting to the default of 6%).

Bug: T383845
Change-Id: I53c3d7e1d7bf6b327388218c9f83c1747bf90df0