knative-serving: add a variable in the templates for the Prometheus port
This is a no-op change to allow to set the Knative Prometheus' port,
rather than hardcode it to 9090. We have some problems with the Calico
network policies, namely no traffic is allowed for port 9090, and
it may be due to the fact that port 9090 is reserved for Calico.

A patch will follow to test a different port for ml-staging-codfw.

Bug: T387580
Change-Id: I6dc33444a7db8b3edc013eabd2f10ef9dfbd36e2