mw-(api-ext|web): shrink next to 80% and increase maxUnavailable
We're seeing issues during deploys that also involve costly image pulls,
as in the case of a new base image.

Since next releases are a bit over-provisioned for the traffic they
receive in practice, scale them down a bit.

Also doubles maxUnavailable, as we're still running with quite a bit of
headroom.

Bug: T383845
Change-Id: Ie9e30b41252ffc8b8ba655399bd23788847e8dd8