mediawiki: Fix envvars with values evaluating to false
Environment variables with values evaluating to false never make it to
the container spec. Check for empty values instead as those are not
supported when php-fpm loads them.

Setting STATSD_EXPORTER_PROMETHEUS_SERVICE_HOST to false is not exactly
the right thing to do, setting it to nil probably is. When set to nil,
it will not be set via container.env but it will be added to the
envvars.inc ConfigMap in order to pass the environment variable
managed by kubernetes (when there is a statsd exporter prometheus
Service object in the namespace) down to php-fpm.

Disable monitoring for wikikube-staging mw-debug deployments as they are
not supposed to be actually used (only to end-to-end test k8s deploys).

Change-Id: Ibb6f172ad4f46952f2c558bea62dca540b2d971e