cirrus: Provide an empty log4j-overrides.properties
It would be nice if we could increase the log levels when
deploying for cases where not everything is working as expected.
This property exists in the crd, but afaict `helmfile --set ...`
will only override existing values, it will not create new nested
values.

As such, add the appropriate value but set it to empty.
Initially this is only applied to cloudelastic for testing, but
should move up to the main values.yaml if it works as expected.

Change-Id: Id9b16185f7c6d84c7eff1fc6ffdf53e8e85d4172