mw-(api-ext|web): right-size given current traffic
This patch does two closely related things:

First, it right-sizes the main and next releases of mw-api-ext and
mw-web given the current fraction of traffic directed to next. It does
so by scaling main and next to 80% and 30% respectively of the multi-DC
serving replica count as if all traffic were served on main.

This is a fairly conservative approach given the current traffic
allocation and utilization of both releases.

Second, it documents anticipated sizes for single-DC serving, using the
projections from https://phabricator.wikimedia.org/P73519 and the same
relative allocations as above. These are close to, but not quite the
same as, the values in Icf079149ab7145df44c50ccccf3e71f9640b8f14 (i.e.,
a couple of replicas higher or lower).

Bug: T383845
Change-Id: I5e68956cb67dda8a9b5e993055a9e39c3c4d73f0