mw-web: additional right-sizing tweaks for next
Follow-on to I5e68956cb67dda8a9b5e993055a9e39c3c4d73f0, which adjusted
sizing of main and next given the observed distribution of traffic.

What that did not account for is that the allocation of traffic to
mw-web/next is a bit more asymmetric across datacenters.

Thus, while keeping replica counts identical across DCs, it makes sense
to add a bit of headroom here to keep worker utilization comparable to
what we're used to.

Even though it's not technically needed - since the asymmetry disappears
in single-DC serving - uses the same allocation (35%) for consistency.

Bug: T383845
Change-Id: Icc8dcfe7149c46984b05117e1481bb836fe07dbe