flink-app: Provide full log4j-console.properties
With some experimentation it's clear that this does not append to
the existing properties, it's a full replacement. As such to
allow an application to have custom logging configuration we need
to copy the log4j-console.properties from flink-kubernetes-operator
over here to flink-app. It seems unfortunate that we now have two
different files with the same expected content, but removing the
configuration from the operator chart also seems incorrect.  Once
we have a working system we can review and hopefully fix as
appropriate.

Change-Id: Ie378cb3b6ef166d45de2ac1f5564268ac91a0e2e