services: Double memory for the kartotherian-main container
It seems that the combination of Nodejs and external C++ libraries
(like Mapnik) is not handled efficiently by the Nodejs' GC.
In particular, the RSS usage grows a lot with traffic, since
the Mapnik objects are not released promptly (eventually they are,
but the process is very slow). There are some workarounds to apply
in the js code to make things more efficient, but the changes
are probably complicated and they would require some time.
I propose to just double the memory allocated by each pod, and observe
some days of traffic to see how the GC behaves with more free
memory (the important bit is not to OOM frequently).

The request involves adding 1.5GiB*40=60GiB of memory for each DC,
something that we definitely can do givent the amount of free RAM
on Wikikube, but of course it needs to be reviewed/approved first
from ServiceOps.

More details in T386926#10617246.

Bug: T386926
Change-Id: Ia1b03ebf75d174faac8d8db6768a1d41ef9b2318