mediawiki-dumps-legacy: deploy the cronjob template in airflow-analytics-test
We plan to run an airflow DAG in charge of
- fetching a cronjob spec
- running a pod of the associated spec, in charge of the dumps

The issue is having the cronjob spec living in another namespace from
the airflow instance, as this would require us to create a ClusterRole
tied to the `aiflow` service account, which would itself require that
the `deploy-airflow` be allowed to create `ClusterRole`, which itself is
a seurity liability.

We are thus going to bring the 2 resources (job spec and airflow) in the
same namespace, which is going to be `airflow-analytics-test` for our
testing phase.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T388378
Change-Id: I3afc03709d08dc4b8bb4f5639515289c9994b81b