mediawiki-dumps-legacy: define a standalone chart for the resources required by the dumps
This chart will contain the Kubernetes resources required by the
mediawiki-dumps-legacy jobs to run, without having to "pollute" the
mediawiki chart itself.

Right now, we define a Role and RoleBinding that should allow the
airflow-analytics/airflow ServiceAccount the required permissions to
fetch the dumps pod spec as well as manage the pods themselves in the
dumps namespace.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T388378
Change-Id: I4f99c2c1930884c8c499e49041341b7828216aed