Move sidecar controller and the pspClusteRole to the mediawiki-dumps-legacy ns
The dumps pods will now run in the mediawiki-dumps-legacy ns, so we need
to reconfigure the sidecar job controller to watch it instead of airflow-analytics-test.

For the same reason, we need to include the allow-mediawiki-psp pspClusterRole to allow
pods with the PTRACE cap and host path volumes to be created in that ns.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T388378
Change-Id: I0c0cbf47becb744cdc01b20575eb985ccdd932cf