services/*: use the correct helm version in each cluster
Use the correct helm version when deploying to a cluster: use helm 3.11
for k8s 1.23 and helm 3.17 for k8s 1.31.

Note that this introduces a trip hazard for mw-* deployments: because
of how the helmfile is structured, kubernetesVersion is hardcoded in
the helmfile for those. Tracked in T388969 and I intend to fix this
ASAP, but in the meantime, this unblocks the k8s upgrade.

Bug: T388390
Change-Id: I48829d3a5c162f7e1d409797d6bdd762d2a3ea86