mw-*: normalize the next and migration releases
This normalizes the structure / content of the values overrides across
all next and migration releases.

In doing so, it also removes specific configuration and mentions of the
PHP 8.1 migration, capturing their more generic use case.

The only functional diff is reverting the maxUnavailable overrides,
while documenting when / why they may be useful.

Bug: T383845
Change-Id: I9e76e578ea64729d1d9a7a8beabda4d6cb9645ae