Allow dot in revision title
Without it, a title with a dot will always fallback to matching the
rest of the path as a title, ignoring the revision.

For example,
https://en.wikipedia.org/api/rest_v1/page/html/Robert_F._Kennedy/1279876094

Errors with,
"The requested relative path
(/v1/page/Robert_F._Kennedy/1279876094/html) did not match any known
handler.

It's unclear why the dot was excluded in 4104a96.

Bug: T389628
Change-Id: Ia671b043b8d1f1cc843ebe46372f95884bfaab90