ingress: Add the ability to have >1 destination rules
Why:
Being able to route traffic between various destinations using a
VirtualService in Istio, requires that DestinationRule entries exist for
each backend service. The current implementation, creates a 1:1 mapping
between VirtualService and DestinationRule which isn't what we want

What:
Add a new version of istio, 1.2.0 which while defaulting to the previous
behavior, allows to define a stanza like

ingress:
  destinationrulehosts:
    - mediawiki-group0-tls-service.mw-api-int.svc.cluster.local
    - mediawiki-group1-tls-service.mw-api-int.svc.cluster.local
    - mediawiki-group2-tls-service.mw-api-int.svc.cluster.local

that allows setting specific DestinationRule records

There is 1 change here for existing charts, that is however a noop.
Existing DestinationRule resources will be renamed. However the name
isn't somehow impacting the functionality, which is why I did not bump
to 2.0.0 and instead opted for 1.2.0

Bug: T384944
Change-Id: Idc9da8b3e60f712818a61bf256da001729895e9e