api-gateway: use rest-gateway for wikifeeds calls to restbase
The feeds endpoints have been deprecated in restbase, and the
api-gateway still uses these at present. Preserve the behaviour by
moving the calls to the rest-gateway. We do this rather than calling
wikifeeds directly in order to centralise accesses to the service and
get better visibility into who is calling the service and how.

Also do some cleanup of labelling, remove restbase references for the
gateway and remove the api-gateway "restbase_dev" option, which was
never used.

Bug: T390317
Change-Id: I2f5cbe6df8d52080d647bd1c4167f8452ae3a590