wikifunctions: Add group{0,1,2} releases
Why:
wikifunctions.org is a wiki running in a dedicated k8s namespace and
releases. This makes it the perfect first wiki for the SingleVersion
HTTP routing. Even though there will be no specific choice to make here
for the backend, we will be sending traffic via the Istio ingressgateway
setup, which is what we really want to showcase here

What:
* Add 3 new releases, using the names group{0,1,2} to mimic the train
* Add the proper values files:
 * No nodePorts for any release
 * group0 defines the ingress stuff, the other 2 releases share it
 * group1 has all the resources assigned to main up to now
 * group2 shouldn't see any traffic whatsoever, minimal resources

Bug: T384944
Change-Id: I009fb9f25c15bf31f39bf0b8a8cebd2127a1a7c4