ml-services: update RRLA prod config
* Add latest image that supports the prediction change event stream
  which was tested on staging as shown in T326179#10681663.

* Add environment variables for this stream.

NB: Unlike the article-country and outlink isvcs where we had to bump
autoscaling resources, the rrla isvc already has high 'minReplicas'
and 'maxReplicas' settings. As Ilias noted, the service seems to be
able to handle many more requests than it currently is. istio shows
12-20 rps and p95 latency is below 300ms. This is ongoing with 5-6
replicas so the assumption is that the current settings are more than
enough.

Bug: T326179
Change-Id: Icc1dae04d61de8106496c25fe4bdb7f82de9bc2b