admin_ng: Preserve Server header in ingressgateway
Why:
Istio leaves envoy at the default configuration regarding HTTP Server
header behavior. This means envoy forcefully sets it to the server_name
stanza, which is "istio-envoy" making it impossible to differentiate
between the backends, which is a feature we want for debugging.

What:
Add an EnvoyFilter configuration, applying it only to the
ingressgateway, that sets APPEND_IF_ABSENT, which will only set the
Server header if the fronted application doesn't set it.  We want to set
this globally for all fronted applications, so the filterChain on
purpose doesn't match on any of the other predicates it supports

Bug: T390854
Bug: T384944
Change-Id: I550784678d5a436aadf6dd97316ee69aa743cdbe