shellbox*: normalize migration releases
This normalizes the structure / content of the values overrides across
all migration releases.

In doing so, it also removes specific configuration and mentions of the
PHP 8.1 migration, capturing their more generic use case.

The only functional diff is down-scaling syntaxhighlight-migration in
staging, which was left running for manual testing purposes.

Bug: T377038
Change-Id: Id9d89cd7b844485316e08d199cc3415b0d0804f5