mesh::configuration: Add sni_rewrites_host_header toggle
Why:
Up to now, we always assumed that if we were setting SNI, we also wanted
to use the SNI value as the HTTP Host header. This made sense for all
the interal services, however in a world of a service being able to
answer to more than 1 virtual hosts, e.g. MediaWiki, this no longer
makes sense.

What:
Add a sni_rewrites_host_header setting per listener, defaulting to true,
which is the current behavior, to allow avoid to set

	auto_host_rewrite: true

The default of Envoy if that value is missing is false.

Change-Id: Ibde4f9850e4714485cd7602576c4a45521b212c7