airflow: set saner performance-related configs
A recent attempt to schedule ~7000 tasks caused the airflow-test-k8s
scheduler to perform poorly. The scheduler was constantly flapping
between healthy and unhealthy, and tasks pods were trickling in
extremely slowly.

This patch tries to set good performance defaults, as well as adds a
metrics on the time the scheduler takes to parse the DAGs.

* We set the number of dag parsing processes to twice the amount of
available vcpus in the scheduler.
* We allow the kubernetes executor to create more than one pod per loop
  (every 5s), which should greatly increase the task throughput
* We increase the max active tasks per tag from 6 to 16 (the airflow
  default)

We also expose configurations such as `scheduler_health_check_threshold`
and `max_tis_per_query` and set the to their default value, to make the
chart a bit more introspectable when it comes to these
peformance-related configs.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T390945
Change-Id: Ib6b2e265d985bc4ab4313e8fc33e9335bf2f4e42