rest-gateway: enable ingress at route-level for citoid
The rest-gateway's config has an uncomfortable distinction between
routes and cluster definitions that parallels Envoy's config but is
somewhat cumbersome for end users. When it comes to enabling ingress
rules we need to inform both route and cluster that we're doing things
to make it work - in this case, setting SNI at the cluster leve, but
also setting the HTTP header at the route level. These actions are
required to be distinct for now, and unless we overhaul the
configuration to more tightly couple route and server definition
(which might be a good idea, but requires a bit of deliberation to get
right), will remain so.

Enable ingress at the route level for citoid.

Bug: T391457
Change-Id: I415d38c8b66ed1c287cf88785892a017484d9be0