airflow: ensure the pod running in the KubernetesPodOperator itself gets low resources
When we're using the KubernetesPodOperator along with the
KubernetesExecutor, we create a pod which only task is to call the
kubernetes API to create a pod and manage its lifetime.

We want to make sure that intermediary pod gets very low resources.
Indeed, in the case of the dumps v1, each dump pod gets a manager pod,
and the request/limit resources really stack up.

Looking at
https://grafana.wikimedia.org/d/-D2KNUEGk/kubernetes-pod-details?orgId=1&var-datasource=eqiad%20prometheus%2Fk8s-dse&var-namespace=airflow-test-k8s&var-pod=mediawiki-sql-xml-regular-dumps-afwiki-dump-metacurren-3enxp4hz&var-container=All&from=1744706723370&to=1744709462003
we can see that that manager pod requires very little resources at all.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T391669
Change-Id: I264338c073ee8e112db791508063bc74380e0c4c