cirrus-streaming-updater: set upgradeMode to savepoint
The last-state approach caused some issues in previous deployment
causing flink to go back to early checkpoints which then all failed
because they were already present on the object store.
Prefer savepoint which is a safer approach and should handle most of the
cases we want to handle from simple restarts to version upgrades with
schema changes.

Bug: T390853
Change-Id: I48c95d3d9fee20f15caae63d18ecb1d51d5d0084