admin_ng: Update Knative on ml-serve-codfw
The ml-serve-codfw cluster is running on Bookworm and containerd,
so it should be ready to progress further into the PSP->PSS migration.

The first step is to upgrade Knative, deploying the new Docker images
but not forcing any new security settings. This should allow us
to deploy kserve-inference services with seccomp profile settings
enabled.

Bug: T369493
Change-Id: Iee5f0bf0f5e93fc6c214ee1c337d9910a3e9a1ad